requirements.txt
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.10.0
reportlab>=4.0.0
bcrypt>=4.0.0
spacy>=3.5.0
cryptography>=41.0.0
click>=8.0.0
pyyaml>=6.0
flask>=3.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
